+++ cmd/dmeta/main.go
@@ -31,6 +31,13 @@ func main() {
 	}
 	addGlazedCommand(rootCmd, "generate-core", generateCore)
 
+	scaffoldInstance, err := dmetacmds.NewScaffoldInstanceCommand()
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "error creating scaffold-instance command: %v\n", err)
+		os.Exit(1)
+	}
+	addGlazedCommand(rootCmd, "scaffold-instance", scaffoldInstance)
+
 	if err := rootCmd.Execute(); err != nil {
 		os.Exit(1)
 	}